+++ app/src/hooks/useParaspellApi.tsx
@@ -8,7 +8,7 @@ import { getSenderAddress } from '@/utils/address'
 import { trackTransferMetrics } from '@/utils/analytics'
 import { isProduction } from '@/utils/env'
 import { extractPapiEvent } from '@/utils/papi'
-import { createRouterPlan } from '@/utils/paraspellSwaps'
+import { createRouterPlan } from '@/utils/paraspellSwap'
 import { createTransferTx, dryRun, DryRunResult, moonbeamTransfer } from '@/utils/paraspellTransfer'
 import { extractPjsEvents } from '@/utils/pjs'
 import { txWasCancelled } from '@/utils/transfer'
